Introducing the attention dimensionality
Previously, attention dimension was hardcoded the same as laten_dim but of course, it could be different. I added the param to the constructor and set its default value to latent_dim. Also, the `create_inital_state` function was doing a bunch of unnecessary calculations while the result was simply discarded. So, I removed the calculations.